spaww mushrooms : done, todo : fix display of bullets

+++ Source/Centiped/Private/CtpPlayerController.cpp
@@ -39,9 +39,9 @@ void ACtpPlayerController::BeginPlay()
 			if (const ACtpGameMode* GameMode = Cast<ACtpGameMode>(GetWorld()->GetAuthGameMode()))
 			{
 				CameraComponent->SetProjectionMode(ECameraProjectionMode::Type::Orthographic);
-				CameraComponent->SetOrthoWidth(GameMode->Width);
+				CameraComponent->SetOrthoWidth(static_cast<float>(GameMode->Width));
 				CameraComponent->SetWorldLocation(FVector(-500, 0, 0));
-				CameraComponent->SetAspectRatio(GameMode->Width / GameMode->Height);
+				CameraComponent->SetAspectRatio(GameMode->Width / static_cast<float>(GameMode->Height));
 
 				CameraActor->SetActorLocation(FVector(-500.f, 0.f, 0.f));
 				CameraActor->SetActorRotation(FRotator(0.f, 0.f, 0.f));
